requirements.txt
requests
google-auth
google-auth-httplib2
google-api-core
google-generativeai
python-dotenv
pyttsx3
sounddevice
scipy
numpy
flask==3.0.0
schedule==1.2.0
llama-cpp-python
pygame
